+++ Tests/Unit/NetworkingDsc.Common.tests.ps1
@@ -532,8 +532,12 @@ try
                     -MockWith { $matchAdapter }
 
                 Mock `
-                    -CommandName Get-ItemPropertyValue `
-                    -MockWith { $noIpv4StaticAddressString }
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            NameServer  = $noIpv4StaticAddressString
+                        }
+                    }
 
                 It 'should not throw exception' {
                     { $script:result = Get-DnsClientServerStaticAddress @ipv4Parameters -Verbose } | Should Not Throw
@@ -545,7 +549,34 @@ try
 
                 It 'should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
-                    Assert-MockCalled -CommandName Get-ItemPropertyValue -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
+                }
+            }
+
+            Context 'Interface Alias was found in system but IPv4 NameServer property does not exist' {
+                Mock `
+                    -CommandName Get-NetAdapter `
+                    -MockWith { $matchAdapter }
+
+                Mock `
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            Dummy = ''
+                        }
+                    }
+
+                It 'should not throw exception' {
+                    { $script:result = Get-DnsClientServerStaticAddress @ipv4Parameters -Verbose } | Should Not Throw
+                }
+
+                It 'should return null' {
+                    $script:result | Should BeNullOrEmpty
+                }
+
+                It 'should call expected mocks' {
+                    Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
                 }
             }
 
@@ -555,8 +586,12 @@ try
                     -MockWith { $matchAdapter }
 
                 Mock `
-                    -CommandName Get-ItemPropertyValue `
-                    -MockWith { $oneIpv4StaticAddressString }
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            NameServer = $oneIpv4StaticAddressString
+                        }
+                    }
 
                 It 'should not throw exception' {
                     { $script:result = Get-DnsClientServerStaticAddress @ipv4Parameters -Verbose } | Should Not Throw
@@ -568,7 +603,7 @@ try
 
                 It 'should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
-                    Assert-MockCalled -CommandName Get-ItemPropertyValue -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
                 }
             }
 
@@ -578,8 +613,12 @@ try
                     -MockWith { $matchAdapter }
 
                 Mock `
-                    -CommandName Get-ItemPropertyValue `
-                    -MockWith { $twoIpv4StaticAddressString }
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            NameServer = $twoIpv4StaticAddressString
+                        }
+                    }
 
                 It 'should not throw exception' {
                     { $script:result = Get-DnsClientServerStaticAddress @ipv4Parameters -Verbose } | Should Not Throw
@@ -592,7 +631,7 @@ try
 
                 It 'should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
-                    Assert-MockCalled -CommandName Get-ItemPropertyValue -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
                 }
             }
 
@@ -602,8 +641,12 @@ try
                     -MockWith { $matchAdapter }
 
                 Mock `
-                    -CommandName Get-ItemPropertyValue `
-                    -MockWith { $noIpv6StaticAddressString }
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            NameServer = $noIpv6StaticAddressString
+                        }
+                    }
 
                 It 'should not throw exception' {
                     { $script:result = Get-DnsClientServerStaticAddress @ipv6Parameters -Verbose } | Should Not Throw
@@ -615,7 +658,34 @@ try
 
                 It 'should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
-                    Assert-MockCalled -CommandName Get-ItemPropertyValue -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
+                }
+            }
+
+            Context 'Interface Alias was found in system but IPv6 NameServer property does not exist' {
+                Mock `
+                    -CommandName Get-NetAdapter `
+                    -MockWith { $matchAdapter }
+
+                Mock `
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            Dummy = ''
+                        }
+                    }
+
+                It 'should not throw exception' {
+                    { $script:result = Get-DnsClientServerStaticAddress @ipv6Parameters -Verbose } | Should Not Throw
+                }
+
+                It 'should return null' {
+                    $script:result | Should BeNullOrEmpty
+                }
+
+                It 'should call expected mocks' {
+                    Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
                 }
             }
 
@@ -625,8 +695,12 @@ try
                     -MockWith { $matchAdapter }
 
                 Mock `
-                    -CommandName Get-ItemPropertyValue `
-                    -MockWith { $oneIpv6StaticAddressString }
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            NameServer = $oneIpv6StaticAddressString
+                        }
+                    }
 
                 It 'should not throw exception' {
                     { $script:result = Get-DnsClientServerStaticAddress @ipv6Parameters -Verbose } | Should Not Throw
@@ -638,7 +712,7 @@ try
 
                 It 'should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
-                    Assert-MockCalled -CommandName Get-ItemPropertyValue -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
                 }
             }
 
@@ -648,8 +722,12 @@ try
                     -MockWith { $matchAdapter }
 
                 Mock `
-                    -CommandName Get-ItemPropertyValue `
-                    -MockWith { $twoIpv6StaticAddressString }
+                    -CommandName Get-ItemProperty `
+                    -MockWith {
+                        [psobject] @{
+                            NameServer = $twoIpv6StaticAddressString
+                        }
+                    }
 
                 It 'should not throw exception' {
                     { $script:result = Get-DnsClientServerStaticAddress @ipv6Parameters -Verbose } | Should Not Throw
@@ -662,7 +740,7 @@ try
 
                 It 'should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
-                    Assert-MockCalled -CommandName Get-ItemPropertyValue -Exactly -Times 1
+                    Assert-MockCalled -CommandName Get-ItemProperty -Exactly -Times 1
                 }
             }
         }